feat: Separate CommandWrap and CommandWrapRef
Make Wrap for deserialize and WrapRef fro serialize.
