+++ src/lib.rs
@@ -1,10 +1,6 @@
 #![cfg_attr(docsrs, feature(doc_auto_cfg))]
 #![doc  = include_str!("../README.md")]
 
-use clap::{Arg, ArgGroup, Command};
-use serde::Deserializer;
-use std::ops::Deref;
-
 #[cfg(not(any(
     feature = "kebab-case-key",
     feature = "snake-case-key",
@@ -23,160 +19,25 @@ compile_error!("Case setting feature is conflicting. Only one should be set.");
 mod de;
 #[cfg(feature = "docsrs")]
 pub mod documents;
+
 #[cfg(feature = "yaml")]
 #[deprecated(since = "0.4", note = "use serde-yaml instead")]
 mod yaml;
+#[cfg(feature = "yaml")]
+pub use yaml::{yaml_to_app, YamlWrap};
 
 mod ser;
+pub use de::app::CommandWrap;
+pub use de::arg::ArgWrap;
 pub use ser::app::NoSkip;
+pub use ser::app::CommandWrapRef;
+pub use ser::arg::ArgWrapRef;
+
+#[deprecated(note = "use CommandWrap instead")]
+pub type AppWrap<'a> = CommandWrap<'a>;
 
 #[cfg(all(test, feature = "snake-case-key"))]
 mod tests;
 
-#[cfg(feature = "yaml")]
-pub use yaml::{yaml_to_app, YamlWrap};
-
-/**
-Deserialize [`Command`] from [`Deserializer`].
-```
-const CLAP_TOML: &'static str = r#"
-name = "app_clap_serde"
-version = "1.0"
-author = "tester"
-about = "test-clap-serde"
-"#;
-let app = clap_serde::load(&mut toml::Deserializer::new(CLAP_TOML))
-    .expect("parse failed");
-assert_eq!(app.get_name(), "app_clap_serde");
-assert_eq!(app.get_about(), Some("test-clap-serde"));
-```
-*/
-pub fn load<'de, D>(de: D) -> Result<Command<'de>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    use serde::Deserialize;
-    CommandWrap::deserialize(de).map(|a| a.into())
-}
-
-/**
-Wrapper of [`Command`] to serialize and deserialize.
-```
-const CLAP_TOML: &'static str = r#"
-name = "app_clap_serde"
-version = "1.0"
-author = "tester"
-about = "test-clap-serde"
-"#;
-let app: clap::Command = toml::from_str::<clap_serde::CommandWrap>(CLAP_TOML)
-    .expect("parse failed")
-    .into();
-assert_eq!(app.get_name(), "app_clap_serde");
-assert_eq!(app.get_about(), Some("test-clap-serde"));
-```
-*/
-#[derive(Debug, Clone)]
-pub struct CommandWrap<'a, S = ()> {
-    app: Command<'a>,
-    ser_setting: S,
-}
-
-impl<'a> CommandWrap<'a> {
-    /// Create a wrapper for [`Command`].
-    pub fn new(app: Command<'a>) -> Self {
-        Self {
-            app,
-            ser_setting: (),
-        }
-    }
-
-    /// Add a setting for serializeing.
-    /// See [`NoSkip`] for details.
-    pub fn with_setting<S>(self, ser_setting: S) -> CommandWrap<'a, S> {
-        CommandWrap {
-            app: self.app,
-            ser_setting,
-        }
-    }
-}
-
-#[deprecated]
-pub type AppWrap<'a> = CommandWrap<'a>;
-
-impl<'a, S> From<CommandWrap<'a, S>> for Command<'a> {
-    fn from(a: CommandWrap<'a, S>) -> Self {
-        a.app
-    }
-}
-
-impl<'a> From<Command<'a>> for CommandWrap<'a> {
-    fn from(app: Command<'a>) -> Self {
-        CommandWrap {
-            app,
-            ser_setting: (),
-        }
-    }
-}
-
-impl<'a> Deref for CommandWrap<'a> {
-    type Target = Command<'a>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.app
-    }
-}
-
-/// Wrapper of [`Arg`] to deserialize with [`DeserializeSeed`](`serde::de::DeserializeSeed`).
-#[derive(Debug, Clone)]
-pub struct ArgWrap<'a, S = ()> {
-    arg: Arg<'a>,
-    pub(crate) ser_setting: S,
-}
-
-impl<'a> ArgWrap<'a> {
-    pub fn new(arg: Arg<'a>) -> Self {
-        Self {
-            arg,
-            ser_setting: (),
-        }
-    }
-}
-
-impl<'a> From<ArgWrap<'a>> for Arg<'a> {
-    fn from(arg: ArgWrap<'a>) -> Self {
-        arg.arg
-    }
-}
-
-impl<'a> From<Arg<'a>> for ArgWrap<'a> {
-    fn from(arg: Arg<'a>) -> Self {
-        ArgWrap {
-            arg,
-            ser_setting: (),
-        }
-    }
-}
-
-impl<'a> Deref for ArgWrap<'a> {
-    type Target = Arg<'a>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.arg
-    }
-}
-
-pub(crate) struct ArgGroupWrap<'a> {
-    group: ArgGroup<'a>,
-}
-
-impl<'a> From<ArgGroupWrap<'a>> for ArgGroup<'a> {
-    fn from(group: ArgGroupWrap<'a>) -> Self {
-        group.group
-    }
-}
-
-impl<'a> From<ArgGroup<'a>> for ArgGroupWrap<'a> {
-    fn from(group: ArgGroup<'a>) -> Self {
-        ArgGroupWrap { group }
-    }
-}
+mod util;
+pub use util::load;
