+++ src/ser/mod.rs
@@ -2,4 +2,4 @@
 mod macros;
 
 pub(crate) mod app;
-mod arg;
+pub(crate) mod arg;
